Create install_ApacheHTTPServer
Let's break down the script:

#!/bin/bash:
This is known as a shebang or hashbang. It specifies the interpreter to be used to execute the script. In this case, it indicates that the script should be interpreted by the Bash shell.

sudo su:
This command attempts to switch the user to the superuser (root). However, since it's in a script, it might not work as expected, especially if it's run non-interactively.

yum update -y:
This command updates all installed packages on the system using the YUM package manager. The -y flag is used to automatically answer "yes" to any prompts, which is useful for unattended installations.

yum install -y httpd.x86_64:
This command installs the Apache HTTP Server (httpd) on the system using YUM. The -y flag is again used to automatically answer "yes" to any prompts.

systemctl start httpd.service:
This command starts the Apache HTTP Server using the systemctl command. It initiates the Apache service.

systemctl enable httpd.service:
This command enables the Apache service to start automatically at boot time.

echo "Hello World from $(hostname -f)" > /var/www/html/index.html:
This command creates a simple HTML file with the content "Hello World from [hostname]" and saves it to the default web server root directory. The $(hostname -f) part is a command substitution that retrieves the fully qualified domain name (FQDN) of the instance.